Created a class for Allure logging removed step from clicking on an article
