+++ src/main/java/org/viesure/utils/AllureLogger.java
@@ -1,0 +1,20 @@
+package org.viesure.utils;
+
+import io.appium.java_client.android.AndroidDriver;
+import io.appium.java_client.android.AndroidElement;
+import io.qameta.allure.Attachment;
+import org.openqa.selenium.OutputType;
+import org.openqa.selenium.TakesScreenshot;
+
+public class AllureLogger {
+
+    @Attachment(value = "Click to open issue screenshot", type = "image/png")
+    public static byte[] saveScreenshotPNG (AndroidDriver<AndroidElement> driver) {
+        return ((TakesScreenshot)driver).getScreenshotAs(OutputType.BYTES);
+    }
+
+    @Attachment(value = "{0}", type = "text/plain")
+    public static String saveTextLog (String message) {
+        return message;
+    }
+}
